+++ packages/firestore/__tests__/firestore.test.ts
@@ -4,6 +4,10 @@ import firestore, {
   firebase,
   Filter,
   getFirestore,
+  getAggregateFromServer,
+  count,
+  average,
+  sum,
   addDoc,
   doc,
   collection,
@@ -650,6 +654,22 @@ describe('Firestore', function () {
 
     it('`enablePersistentCacheIndexAutoCreation` is properly exposed to end user', function () {
       expect(enablePersistentCacheIndexAutoCreation).toBeDefined();
+    });
+
+    it('`getAggregateFromServer` is properly exposed to end user', function () {
+      expect(getAggregateFromServer).toBeDefined();
+    });
+
+    it('`count` is properly exposed to end user', function () {
+      expect(count).toBeDefined();
+    });
+
+    it('`average` is properly exposed to end user', function () {
+      expect(average).toBeDefined();
+    });
+
+    it('`sum` is properly exposed to end user', function () {
+      expect(sum).toBeDefined();
     });
   });